+++ tests/testthat/test-data.R
@@ -13,14 +13,14 @@ test_that("Component construction: default mesh/mapping, data is list", {
   )
 
   cmp1 <- bru_component_list(~ effect(c(1, 1.5, 2, 3, 4), model = "iid") - 1)
-  cmp2 <- add_mappers(cmp1, lhoods = bru_like_list(list(lik)))
+  cmp2 <- add_mappers(cmp1, lhoods = bru_obs_list(list(lik)))
   expect_equal(
     ibm_values(cmp2$effect$mapper, multi = 1)$main,
     sort(unique(lik$data$x), na.last = NA)
   )
 
   cmp1 <- bru_component_list(~ effect(x, model = "rw2") - 1)
-  cmp2 <- add_mappers(cmp1, lhoods = bru_like_list(list(lik)))
+  cmp2 <- add_mappers(cmp1, lhoods = bru_obs_list(list(lik)))
   expect_equal(
     ibm_values(cmp2$effect$mapper, multi = 1)$main,
     sort(unique(lik$data$x), na.last = NA)
@@ -42,7 +42,7 @@ test_that("Component construction: default mesh/mapping, data is list", {
       mapper = bru_mapper(mesh1, indexed = FALSE)
     ) - 1
   )
-  cmp2 <- add_mappers(cmp1, lhoods = bru_like_list(list(lik)))
+  cmp2 <- add_mappers(cmp1, lhoods = bru_obs_list(list(lik)))
   expect_equal(
     ibm_values(cmp2$effect$mapper, multi = 1)$main,
     sort(unique(lik$data$x), na.last = NA)
@@ -54,7 +54,7 @@ test_that("Component construction: default mesh/mapping, data is list", {
       mapper = bru_mapper(mesh1, indexed = TRUE)
     ) - 1
   )
-  cmp2 <- add_mappers(cmp1, lhoods = bru_like_list(list(lik)))
+  cmp2 <- add_mappers(cmp1, lhoods = bru_obs_list(list(lik)))
   expect_equal(
     ibm_values(cmp2$effect$mapper, multi = 1)$main,
     seq_along(sort(unique(lik$data$x), na.last = NA))
@@ -74,7 +74,7 @@ test_that("Component construction: unsafe intercepts, data is list", {
   lik <- bru_used_update(lik, labels = names(cmp))
   expect_error(
     object = {
-      model <- bru_model(cmp, bru_like_list(list(lik)))
+      model <- bru_model(cmp, bru_obs_list(list(lik)))
     },
     paste0(
       "The input evaluation 'something_unknown' for 'something_unknown' ",
@@ -111,7 +111,7 @@ test_that("Component construction: separate response_data input", {
 
 
   cmp1 <- bru_component_list(~ effect(x, model = "rw2") - 1)
-  cmp2 <- add_mappers(cmp1, lhoods = bru_like_list(list(lik2)))
+  cmp2 <- add_mappers(cmp1, lhoods = bru_obs_list(list(lik2)))
   expect_equal(ibm_values(cmp2$effect$mapper, multi = 1)$main, lik2$data$x)
 
   mesh1 <- fm_mesh_1d(lik1$data$x)
@@ -128,7 +128,7 @@ test_that("Component construction: separate response_data input", {
       mapper = bru_mapper(mesh1, indexed = FALSE)
     ) - 1
   )
-  cmp2 <- add_mappers(cmp1, lhoods = bru_like_list(list(lik1)))
+  cmp2 <- add_mappers(cmp1, lhoods = bru_obs_list(list(lik1)))
   expect_equal(ibm_values(cmp2$effect$mapper, multi = 1)$main, lik1$data$x)
 
   cmp1 <- bru_component_list(
@@ -137,12 +137,12 @@ test_that("Component construction: separate response_data input", {
       mapper = bru_mapper(mesh1, indexed = TRUE)
     ) - 1
   )
-  cmp2 <- add_mappers(cmp1, lhoods = bru_like_list(list(lik1)))
+  cmp2 <- add_mappers(cmp1, lhoods = bru_obs_list(list(lik1)))
   expect_equal(
     ibm_values(cmp2$effect$mapper, multi = 1)$main,
     seq_along(lik1$data$x)
   )
-  cmp2 <- add_mappers(cmp1, lhoods = bru_like_list(list(lik2)))
+  cmp2 <- add_mappers(cmp1, lhoods = bru_obs_list(list(lik2)))
   expect_equal(
     ibm_values(cmp2$effect$mapper, multi = 1)$main,
     seq_along(lik2$data$x)