+++ tests/testthat/test-linearisation.R
@@ -11,7 +11,7 @@ test_that("Linearisation", {
 
   cmp <- ~ -1 + x + Int_y(1) + Int_z(1)
   lhoods <-
-    bru_like_list(
+    c(
       bru_obs(
         formula = y ~ exp(x) + Int_y_latent,
         data = data
@@ -37,8 +37,9 @@ test_that("Linearisation", {
   expect_equal(used[["latent"]], c("Int_y", "Int_z"))
 
   model <- bru_model(bru_component_list(cmp), lhoods)
+  lhoods <- model$lhoods
 
-  idx <- evaluate_index(model, lhoods)
+  idx <- evaluate_index(model, used = bru_used(lhoods))
   inp <- evaluate_inputs(model, lhoods)
   comp_lin <- evaluate_comp_lin(model, input = inp, state = NULL)
   lin0 <- bru_compute_linearisation.bru_model(
@@ -114,7 +115,7 @@ test_that("Linearisation", {
       c(stks0, list(compress = TRUE, remove.unused = FALSE))
     )
 
-  stk0_ <- bru_make_stack.bru_like_list(lhoods, lin0, idx)
+  stk0_ <- bru_make_stack(lhoods, lin0, idx)
 
   expect_s3_class(stk0, "inla.data.stack")